refactor: change Float and Double types to double in TopSpendingCategory

+++ lib/domain/repository/model/top_spending_category.dart
@@ -5,9 +5,9 @@ import 'package:moneyplus/domain/entity/transaction_category.dart';
 class TopSpendingCategory {
   final TransactionCategory category;
   final int numberOfTransactions;
-  final Double total;
+  final double total;
   final String currency;
-  final Float percentage;
+  final double percentage;
 
   TopSpendingCategory({
     required this.category,
